requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-bootstrap-breadcrumbs==0.9.1
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-disqus==0.5
django-hitcount==1.3.0
django-jet==1.0.8
django-js-asset==1.1.0
django-taggit==0.24.0
django-tracking2==0.5.0
django-view-breadcrumbs==0.0.8
djangorestframework==3.9.1
google-api-python-client==1.4.1
httplib2==0.12.0
idna==2.8
isort==4.3.4
jet-django==0.4.6
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
oauth2client==4.1.3
pbr==5.1.3
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
